+++ CHANGES.md
@@ -1,12 +1,12 @@
 # Version 0.9.6
 
-- Added issue templates for GitHub
-
 - Added interactive plot for connectivity (iplotConn)
 
 - Added 'dynamicRates' option for NetStim populations
 
 - Added option to have a uniform distribution of rates for Vecstim pulses
+
+- Added function for distributed saving at intervals
 
 - Added issue templates for GitHub